+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Added "Flow Matching" examples/demo.
 * Added `layers.DropBlock`, a type of dropout for images.
-* Added `layers.DropPath`, a type of dropout used in Residual connections, to drop full paths.
+* Added `layers.DropPath` and `layers.DropPathFromContext`, a type of dropout used in Residual connections, to drop full paths.
 * Added `Context.RandomBenoulli` to sample from a Bernoulli (binary) distribution.
 * Correctly pretty-print Float16 and BFloat16 tensors.
 * Fixed nanlogger for Float16 and BFloat16; Also, it first prints other logged tensors, before failing with a NaN.
@@ -12,6 +12,8 @@
 * `nanlogger`:
   * Store only the stack-trace, and trim the stack into the nanlogger package.
   * Does not exit, simply report the NanLogger. User can define a handler, if they want the training to exit.
+* `layers.LayerNormalization`: up-scale precision by default if input is a Float16 or BFloat16. Low-precision
+  lead to NaNs when reducing values for normalization. Added also a hyperparameter to configure normalization DType.
 
 # v0.16.1 - 🎄 2024/12/19 🎄 MatMul fixes
 * MatMul fixed for some edge shape configuration and greatly accelerated in some cases.